requirements.txt
flask
scikit-learn==1.2.2
python-whois